Moved logo + providers to separate shared directory
