+++ frontend/src/components/shared/Logo.tsx
